+++ src/App.js
@@ -10,6 +10,7 @@ import Anonym from "./page/Community/Anonym/Anonym";
 import Feed from "./page/Feed/Feed";
 import Roadview from "./page/Roadview/Roadview";
 import Schoolschdule from "./page/Schoolschdule/Schoolschdule";
+import Timetable from "./page/Timetable/Timetable";
 function App() {
   return (
     <BrowserRouter>
@@ -23,6 +24,7 @@ function App() {
           <Route path='/welbato/community' element={<Normal/>}/>
           <Route path='/welbato/feed' element={<Feed/>}/>
           <Route path='/welbato/roadview' element={<Roadview/>}/>
+          <Route path='/welbato/schdule' element={<Timetable/>}/>
           <Route path='/welbato/schoolschdule' element={<Schoolschdule/>}/>
         </Routes>
     </BrowserRouter>
