+++ src/page/Timetable/Timetable.scss
@@ -1,0 +1,131 @@
+@import '../../style/common.scss';
+
+.Time-section {
+    height: 100vh;
+    background-color: $background-blue;
+}
+
+.Time-tit {
+    height: 100%;
+    background-color: #EFF5FF;
+
+    .lunch-top {
+        position: relative;
+        top: 90px;
+        left: 30px;
+        display: inline-block;
+
+        .lunch-top-tit {
+            .annaemon {
+                float: left;
+
+                img {}
+            }
+
+            span {
+                display: inline-block;
+                position: relative;
+                top: 10px;
+                left: 10px;
+
+                p {
+                    font-weight: lighter;
+                    margin: 0 auto;
+                }
+            }
+        }
+    }
+
+    .lunch-bottom {
+        position: fixed;
+        bottom: 0px;
+        width: 100%;
+        height: 70%;
+
+        .lunch-bottom-tit {
+            .lunch-day {
+                text-align: center;
+                position: relative;
+                top: 10px;
+
+                span {
+                    font-style: normal;
+                    font-weight: 400;
+                    font-size: 25px;
+                }
+            }
+
+            .lunch-bottom-bottom {
+                display: flex;
+                justify-content: center;
+                position: relative;
+                top: 40px;
+
+                .lunch-bottom-time {
+                    width: 20%;
+
+                    .lunch-time {
+                        display: flex;
+                        justify-content: space-between;
+
+                        button {
+                            background-color: transparent;
+                            border: none;
+
+                            img {}
+
+                            &:hover {
+                                cursor: pointer;
+                            }
+                        }
+                    }
+                }
+            }
+
+            .lunch-bottom-menu {
+                width: 100%;
+                position: relative;
+                top: 70px;
+                display: flex;
+                justify-content: center;
+                .lunch-bottom-menu-musa {
+                    align-items: center;
+                    width: 45%;
+                    display: flex;
+                    justify-content: space-between;
+                    .lunch-bottom-menu-button {
+                        button {
+                            background-color: white;
+                            border: none;
+                            border-radius: 130px;
+                            width: 65px;
+                            height: 65px;
+                            img {
+                                width: 18px;
+                            }
+                            &:hover{
+                                cursor: pointer;
+                            }
+                        }
+                    }
+
+                    .lunch-bottom-menu-menu {
+                        align-items: center;
+                        text-align: center;
+                        background-color: white;
+                        width: 60%;
+                        height: 300px;
+                        border-radius: 10px;
+                        p {
+                            
+                            span {
+                                font-size: 18px;
+                            }
+                        }
+                    }
+
+                }
+            }
+        }
+    }
+}
